Wppl binary prints to temp file; closes #2
`printWebPPLValue` was modified to print to `/tmp/webppl_output`.
`system2` call from R no longer waits for webppl to finish, and so
prints to the RConsole instead of capturing the output.
After the system call, R waits until the file exists and then reads it.

This will fail if the output from WebPPL takes a long time to write to file.
The file will begin existing (but WebPPL won't be done printing), and R will read it.