Refactor tidy_output for wppl >0.9.6; tests for output format
I introduced a number of local functions to rwebppl.R, to test for what kind
of output format is entering into R. This is indepent of the output_format
specified in the webppl call from R. I pushed those conditions back in the
queue of if/else statements. It's likely those can either be gotten rid of, or
will need to be updated.