Remove output_format option; clean ggmcmc/tidy format; update docs & tests
As of webppl 0.9.6, sampling algorithms return samples automatically, obviating the need for the `output_format="samples"` option (already, when available, tidy_output will render samples).
Now I learned that the order in which the samples are returned is,
in fact, the order in which they were collected,
so we have the correspondence between iteration number of the MCMC chain and the sample.
This change makes it so, for sampling algorithms, we default to the output format currently called `"ggmcmc"`.
This format is long-form, with columns: `Iteration | Parameter | value | Chain` and can be used out of the box with the `ggmcmc` package.

The tidy/ggmcmc format requires some meta data, which we take from inference_opts argument when available.
When inference options are specified inside webppl, we can look to the length of the samples list for the
number of samples.
Burn and thin parameters are currently unavailable, but may be added with [WebPPL issue #608](https://github.com/probmods/webppl/issues/608#issuecomment-278721818).

Tests and docs are updated accordingly.